Major rework of the sbt-dotty plugin for cross-compilation support
The plugin can now be used to cross-compile between Scala 2 and Dotty:
Just enabling the plugin will not change anything in your build, you
need to set scalaVersion to a Dotty version (starting with "0.") to
enable Dotty support. This also works with `crossScalaVersions`.

The plugin now works with the newly-published bootstrapped version of
dotty

The plugin now uses `scalaBinaryVersion` 0.1 instead of 2.11, dotty
still supports depending on 2.11 artifacts but you need to use
`withDottyCompat` (added in the next commit).